dancing_devil
This project is used to demonstrate the creation of a devil and its rotations in all the three axis as described below.
The devil is constructed or divided into head, body , hands, eyes, horns , tail and stick. The stick is further divided into upper portion of the stick, lower portion of the stick, sphere on the top of the stick and sharp spikes has been created for its stick.
Sharp spikes has been added to the stick so as to make it more attractive. A ring
has been placed on the devils head. Further a plate has been created which appears as though the devil is standing on it.
The enhancements for this project are that we can provide movement for hand, legs and stick of the devil. Further these types can be used in the animation industry. Further by providing lighting and shadows to the object it make it more attractive.